requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
asttokens==2.4.1
attrs==23.2.0
blinker==1.8.2
cachetools==5.3.3
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
cryptography==42.0.8
dataclasses-json==0.6.7
debugpy==1.8.2
decorator==5.1.1
docstring_parser==0.16
executing==2.0.1
Flask==3.0.3
frozenlist==1.4.1
google-api-core==2.19.1
google-auth==2.32.0
google-cloud-aiplatform==1.49.0
google-cloud-appengine-logging==1.4.4
google-cloud-audit-log==0.2.5
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
google-cloud-logging==3.10.0
google-cloud-resource-manager==1.12.4
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
greenlet==3.0.3
grpc-google-iam-v1==0.13.1
grpcio==1.65.0
grpcio-status==1.62.2
gunicorn==22.0.0
idna==3.7
ipykernel==6.29.5
ipython==8.26.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
langchain==0.2.7
langchain-community==0.2.7
langchain-core==0.2.16
langchain-text-splitters==0.2.2
langsmith==0.1.85
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib-inline==0.1.7
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==1.26.4
orjson==3.10.6
packaging==24.1
parso==0.8.4
pdfminer.six==20240706
pexpect==4.9.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==4.25.3
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
PyMuPDF==1.24.7
PyMuPDFb==1.24.6
pypdf==4.2.0
python-dateutil==2.9.0.post0
PyYAML==6.0.1
pyzmq==26.0.3
referencing==0.35.1
requests==2.32.3
rpds-py==0.19.0
rsa==4.9
shapely==2.0.4
six==1.16.0
SQLAlchemy==2.0.31
stack-data==0.6.3
tenacity==8.5.0
tornado==6.4.1
traitlets==5.14.3
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.2
vertexai==1.49.0
wcwidth==0.2.13
Werkzeug==3.0.3
yarl==1.9.4
